LayoutTests/fast/css-grid-layout/grid-element-repeat-get-set-expected.txt

Test that setting and getting grid-template-columns and grid-template-rows with repeat() works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting grid-template-columns and grid-template-rows set through CSS
PASS window.getComputedStyle(singleSingleTrackRepeatWithSize, '').getPropertyValue('-webkit-grid-template-columns') is "120px"
PASS window.getComputedStyle(singleSingleTrackRepeatWithSize, '').getPropertyValue('-webkit-grid-template-rows') is "18px"
PASS window.getComputedStyle(singleSingleTrackRepeatWithoutSize, '').getPropertyValue('-webkit-grid-template-columns') is "0px"
PASS window.getComputedStyle(singleSingleTrackRepeatWithoutSize, '').getPropertyValue('-webkit-grid-template-rows') is "18px"
PASS window.getComputedStyle(twoSingleTrackRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "400px 400px"
PASS window.getComputedStyle(twoSingleTrackRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "0px 0px"
PASS window.getComputedStyle(twoSingleTrackRepeatWithChildren, '').getPropertyValue('-webkit-grid-template-columns') is "400px 400px"
PASS window.getComputedStyle(twoSingleTrackRepeatWithChildren, '').getPropertyValue('-webkit-grid-template-rows') is "44px 77px"
PASS window.getComputedStyle(twoDoubleTrackRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "33px 120px 0px 120px"
PASS window.getComputedStyle(twoDoubleTrackRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "10px 77px 10px 0px"
PASS window.getComputedStyle(twoDoubleTrackWithNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "33px middle 250px end 0px middle 250px end"
PASS window.getComputedStyle(twoDoubleTrackWithNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "10px start 77px end 10px start 0px end"
PASS window.getComputedStyle(twoDoubleTrackWithTrailingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "before 0px before 0px"
PASS window.getComputedStyle(twoDoubleTrackWithTrailingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "before 10px before 10px"
PASS window.getComputedStyle(trailingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "250px end"
PASS window.getComputedStyle(trailingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "10px end"
PASS window.getComputedStyle(leadingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "start 250px 250px"
PASS window.getComputedStyle(leadingNamedGridLineRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "start 10px 10px"
PASS window.getComputedStyle(mixRepeatAfterNonRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "start 140px 250px 250px"
PASS window.getComputedStyle(mixRepeatAfterNonRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "44px 10px 10px"
PASS window.getComputedStyle(mixNonRepeatAfterRepeat, '').getPropertyValue('-webkit-grid-template-columns') is "250px 250px 120px last"
PASS window.getComputedStyle(mixNonRepeatAfterRepeat, '').getPropertyValue('-webkit-grid-template-rows') is "10px 10px end 0px"

Test invalid repeat syntax.
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-template-columns') is "none"
PASS successfullyParsed is true

TEST COMPLETE
